+++ R/register_all_transpilers.R
@@ -8,6 +8,7 @@ register_all_transpilers <- function() {
   append_builtin_transpilers_for_future.apply()
   append_builtin_transpilers_for_purrr()
   append_builtin_transpilers_for_furrr()
+  append_builtin_transpilers_for_partykit()
   append_builtin_transpilers_for_plyr()
   append_builtin_transpilers_for_stats()
   append_builtin_transpilers_for_foreach()